fix(YouTube/Spoof signature): Removed from settings, integrated into `GmsCore support` patch
