+++ app/src/main/java/app/revanced/integrations/shared/settings/BaseSettings.java
@@ -23,6 +23,4 @@ public class BaseSettings {
      */
     public static final BooleanSetting HIDE_FULLSCREEN_ADS = new BooleanSetting("revanced_hide_fullscreen_ads", TRUE, true);
     public static final BooleanSetting HIDE_FULLSCREEN_ADS_TYPE = new BooleanSetting("revanced_hide_fullscreen_ads_type", TRUE, true);
-
-    public static final BooleanSetting SPOOF_SIGNATURE = new BooleanSetting("revanced_spoof_signature", FALSE, true, "revanced_spoof_signature_user_dialog_message");
 }
